feat: add Collapsible component and IconSymbol updates for improved icon handling and UI consistency
