+++ app/(tabs)/explore.tsx
@@ -2,12 +2,12 @@ import { Image } from 'expo-image';
 import { Platform, StyleSheet } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
+import Collapsible from '@/components/common/Collapsible';
+import IconSymbol from '@/components/common/IconSymbol';
 import ThemedText from '@/components/common/ThemedText';
 import ThemedView from '@/components/common/ThemedView';
 import ExternalLink from '@/components/ExternalLink';
 import ParallaxScrollView from '@/components/ParallaxScrollView';
-import Collapsible from '@/components/ui/Collapsible';
-import IconSymbol from '@/components/ui/IconSymbol';
 import { useTheme } from '@/hooks/useTheme';
 import { commonStyles } from '@/styles/common';
 import { fonts } from '@/theme';
