+++ components/common/IconSymbol/constants.ts
@@ -1,0 +1,19 @@
+import MaterialIcons from '@expo/vector-icons/MaterialIcons';
+import { SymbolViewProps } from 'expo-symbols';
+import { ComponentProps } from 'react';
+
+export type IconMapping = Record<SymbolViewProps['name'], ComponentProps<typeof MaterialIcons>['name']>;
+
+/**
+ * Add your SF Symbols to Material Icons mappings here.
+ * - see Material Icons in the [Icons Directory](https://icons.expo.fyi).
+ * - see SF Symbols in the [SF Symbols](https://developer.apple.com/sf-symbols/) app.
+ */
+export const ICON_MAPPING = {
+  'house.fill': 'home',
+  'paperplane.fill': 'send',
+  'chevron.left.forwardslash.chevron.right': 'code',
+  'chevron.right': 'chevron-right',
+} as IconMapping;
+
+export type IconSymbolName = keyof typeof ICON_MAPPING;
