+++ components/common/IconSymbol/index.ios.tsx
@@ -1,5 +1,13 @@
 import { SymbolView, SymbolViewProps, SymbolWeight } from 'expo-symbols';
-import { StyleProp, ViewStyle } from 'react-native';
+import { type StyleProp, type ViewStyle } from 'react-native';
+
+type Props = {
+  name: SymbolViewProps['name'];
+  size?: number;
+  color: string;
+  style?: StyleProp<ViewStyle>;
+  weight?: SymbolWeight;
+};
 
 const IconSymbol = ({
   name,
@@ -7,13 +15,7 @@ const IconSymbol = ({
   color,
   style,
   weight = 'regular',
-}: {
-  name: SymbolViewProps['name'];
-  size?: number;
-  color: string;
-  style?: StyleProp<ViewStyle>;
-  weight?: SymbolWeight;
-}) => {
+}: Props) => {
   return (
     <SymbolView
       weight={weight}
